fix(product): split hero exit overlay out of the hooks layer
The frontend layer boundary checker rejects hooks/ -> components/ imports.
use-chat-loading-hero-exit.tsx mixed the pure exit-hold hook with a JSX
overlay that pulled in ChatPreMessageCanvas/WorkspaceCreationReceipt. Split
into a pure hook (hooks/chat/ui/use-chat-loading-hero-exit.ts) and a
component (components/workspace/chat/surface/ChatLoadingHeroExitOverlay.tsx)
that imports the hook's exported phase type. No behavior change.
